+++ streamx-flink/streamx-flink-repl/src/main/scala/com/streamxhub/repl/flink/interpreter/FlinkScalaInterpreter.scala
@@ -44,7 +44,7 @@ import org.apache.flink.streaming.api.TimeCharacteristic
 import org.apache.flink.streaming.api.environment.{StreamExecutionEnvironmentFactory, StreamExecutionEnvironment => JStreamExecutionEnvironment}
 import org.apache.flink.streaming.api.scala.StreamExecutionEnvironment
 import org.apache.flink.yarn.entrypoint.YarnJobClusterEntrypoint
-import org.apache.zeppelin.interpreter.{InterpreterContext}
+import org.apache.zeppelin.interpreter.InterpreterContext
 
 import scala.tools.nsc.interpreter.IR
 import org.slf4j.LoggerFactory
